ai-service/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
openai>=1.3.0
google-generativeai>=0.3.0
requests>=2.31.0
pydantic>=2.0.0
python-multipart>=0.0.6
celery>=5.3.0
redis>=5.0.0
httpx>=0.25.0
transformers>=4.36.0
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
aiofiles>=23.2.0
